Add close definitions for Response/Request handles
This commit takes the changes from the internal repo and exposes them
ahead of our new SDK release so that we can also update viceroy before
then to be able to also handle the fact that handles have a new drop
implementation for the Rust SDK that will automatically close themselves
out when they go out of scope. If people use Viceroy with the new SDK
version, then their programs will constantly crash and we can't have
that happen.