Add auto_decompress_response_set host call (#7)
With the newest 0.8.1 SDK release we need to move the host call to the
public repo so that we can update Viceroy and fix
https://github.com/fastly/Viceroy/issues/110. We just need to stub out
the hostcall there, but it depends on this repo to have the latest
definitions. This commit promotes the hostcall from an internal one to
an external.